# fix: optimize workflow runs by returning early if there are no workflows