--- a/app/controlplane/pkg/data/workflowrun.go
+++ b/app/controlplane/pkg/data/workflowrun.go
@@ -25,7 +25,6 @@ import (
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/biz"
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/data/ent"
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/data/ent/attestation"
-	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/data/ent/organization"
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/data/ent/projectversion"
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/data/ent/workflow"
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/data/ent/workflowrun"
@@ -234,10 +233,24 @@ func (r *WorkflowRunRepo) List(ctx context.Context, orgID uuid.UUID, filters *bi
 		return nil, "", errors.New("pagination options is required")
 	}
 
+	// query first for workflows to avoid joining the workflow_runs table
+	wfExist, err := r.data.DB.Workflow.Query().Where(
+		workflow.DeletedAtIsNil(),
+		workflow.OrganizationID(orgID),
+	).Exist(ctx)
+	if err != nil {
+		return nil, "", fmt.Errorf("getting workflows: %w", err)
+	}
+	if !wfExist {
+		// No workflows in the org, no runs
+		return nil, "", nil
+	}
+
+	// Query workflow runs by joining with workflows
 	q := r.data.DB.WorkflowRun.Query().Where(
 		workflowrun.HasWorkflowWith(
+			workflow.OrganizationID(orgID),
 			workflow.DeletedAtIsNil(),
-			workflow.HasOrganizationWith(organization.ID(orgID)),
 		)).
 		Order(ent.Desc(workflowrun.FieldCreatedAt)).
 		WithWorkflowAndProject().WithVersion().